requirements.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.6
django-jet==1.0.8
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
idna==3.2
mysqlclient==2.0.3
pycodestyle==2.7.0
PyJWT==2.1.0
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
stripe==2.60.0
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.6
